win32: fix functional keys, native caption glyphs, maximized blur, initial pty size
- keycodes[] widened from short to int so FKEY values (>=0xe000) no longer
  sign-extend to 0xffffXXXX. Enter/Backspace/Tab/arrows now encode correctly
  and reach the child; this also lets right-arrow/tab accept PSReadLine
  inline predictions.
- rework WM_KEYDOWN/WM_CHAR: forward functional keys and shortcut-modified
  keys via keydown, deliver plain text via WM_CHAR, and suppress the
  duplicate control-char WM_CHAR.
- caption buttons render the native Segoe Fluent Icons / MDL2 Assets glyphs
  (E921/E922/E923/E8BB) instead of hand-drawn shapes.
- acrylic blur falls back to classic blur-behind when maximized (DWM disables
  acrylic for zoomed windows) and is re-applied across maximize/restore.
- create the ConPTY at the real window size instead of a fixed 80x24 so the
  shell first prompt is laid out correctly.
- drop the "WxH cells" live-resize banner on Windows.

+++ glfw/win32_platform.h
@@ -69,7 +69,7 @@ typedef struct _GLFWlibraryWin32 {
     HDEVNOTIFY   deviceNotificationHandle;
     int          acquiredMonitorCount;
     char*        clipboardString;
-    short int    keycodes[512];
+    int          keycodes[512];
     short int    scancodes[512];
     char         keynames[512][5];
